refactor : 도메인에 있는 로직 분리
