+++ src/test/kotlin/lotto/model/LottoWinnersTest.kt
@@ -2,7 +2,6 @@ package lotto.model
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import lotto.view.incomeStatement
 
 class LottoWinnersTest : StringSpec({
 
@@ -40,20 +39,5 @@ class LottoWinnersTest : StringSpec({
         val boom = LottoWinners(totalGameCount = 20)
         val actual = boom.earningRate(Issuer.pricePerGame())
         actual shouldBe 0.toDouble()
-    }
-
-    "0일때는 손해로 출력 되어야 한다" {
-        val actual = (0.00).toDouble().incomeStatement()
-        actual shouldBe "손해"
-    }
-
-    "1일 미만인 경우 손해로 출력 되어야 한다" {
-        val actual = (0.99).incomeStatement()
-        actual shouldBe "손해"
-    }
-
-    "1 일때는 이익으로 출력 되어야 한다" {
-        val actual = (1.00).toDouble().incomeStatement()
-        actual shouldBe "이익"
     }
 })
